+++ app/src/androidTest/java/com/company/mobile/android/appname/app/browse/BrowseBufferoosViewModelTest.kt
@@ -6,8 +6,8 @@ import com.company.mobile.android.appname.app.browse.BrowseState.Loading
 import com.company.mobile.android.appname.app.browse.BrowseState.Success
 import com.company.mobile.android.appname.app.test.util.BufferooFactory
 import com.company.mobile.android.appname.app.test.util.DataFactory
-import com.company.mobile.android.appname.data.browse.Bufferoo
-import com.company.mobile.android.appname.data.browse.interactor.GetBufferoos
+import com.company.mobile.android.model.bufferoo.Bufferoo
+import com.company.mobile.android.appname.domain.bufferoo.interactor.GetBufferoos
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever