+++ data/src/main/java/com/company/mobile/android/appname/data/browse/Bufferoo.kt
@@ -1,6 +1,0 @@
-package com.company.mobile.android.appname.data.browse
-
-/**
- * Representation for a [Bufferoo] fetched from an external layer data source
- */
-data class Bufferoo(val id: Long, val name: String, val title: String, val avatar: String)