+++ domain/src/main/java/com/company/mobile/android/appname/domain/bufferoo/repository/BufferooRepository.kt
@@ -1,0 +1,14 @@
+package com.company.mobile.android.appname.domain.bufferoo.repository
+
+import com.company.mobile.android.model.bufferoo.Bufferoo
+import io.reactivex.Completable
+import io.reactivex.Single
+
+interface BufferooRepository {
+
+    fun clearBufferoos(): Completable
+
+    fun saveBufferoos(bufferoos: List<Bufferoo>): Completable
+
+    fun getBufferoos(): Single<List<Bufferoo>>
+}